+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Improve remote server join time
 - Fix sound volumes not caculating properly across multiple maps
 - Fix storage slot saving not working properly
+- Fix sounds not clearing properly when hiding instances
 
 ## [0.7.7] 2026-04-03